Refactored library for easier patching
* Moved specific patches for builtin prototypes to `/patches`
* refactored existing code into smaller modules:
 - style-scoping
 - link-nodes
 - patch-(instances, prototypes, native, shadyroot)
