+++ src/patches/Slot.js
@@ -1,0 +1,31 @@
+/**
+@license
+Copyright (c) 2016 The Polymer Project Authors. All rights reserved.
+This code may only be used under the BSD style license found at http://polymer.github.io/LICENSE.txt
+The complete set of authors may be found at http://polymer.github.io/AUTHORS.txt
+The complete set of contributors may be found at http://polymer.github.io/CONTRIBUTORS.txt
+Code distributed by Google as part of the polymer project is also
+subject to an additional IP rights grant found at http://polymer.github.io/PATENTS.txt
+*/
+
+import {shadyDataForNode} from '../shady-data.js';
+import {renderRootNode} from './Slotable.js';
+
+export const Slot = {
+
+  /**
+   * @this {HTMLSlotElement}
+   * @param {object} options
+   */
+  assignedNodes(options) {
+    if (this.localName === 'slot') {
+      renderRootNode(this);
+      const nodeData = shadyDataForNode(this);
+      return nodeData ?
+        ((options && options.flatten ? nodeData.flattenedNodes :
+          nodeData.assignedNodes) || []) :
+        [];
+    }
+  }
+
+};
