+++ src/patches/Window.js
@@ -1,0 +1,23 @@
+/**
+@license
+Copyright (c) 2016 The Polymer Project Authors. All rights reserved.
+This code may only be used under the BSD style license found at http://polymer.github.io/LICENSE.txt
+The complete set of authors may be found at http://polymer.github.io/AUTHORS.txt
+The complete set of contributors may be found at http://polymer.github.io/CONTRIBUTORS.txt
+Code distributed by Google as part of the polymer project is also
+subject to an additional IP rights grant found at http://polymer.github.io/PATENTS.txt
+*/
+
+import {addEventListener, removeEventListener} from '../patch-events.js';
+
+// TODO(sorvell) unsure if this is needed now that we patch EventTarget.
+export const Window = {
+
+  // NOTE: ensure these methods are bound to `window` so that `this` is correct
+  // when called directly from global context without a receiver; e.g.
+  // `addEventListener(...)`.
+  addEventListener: addEventListener.bind(window),
+
+  removeEventListener: removeEventListener.bind(window)
+
+};
